+++ src/components/Header.js
@@ -1,0 +1,10 @@
+function Header(){
+    return(
+        <header>
+          <h1>LabeTube</h1>
+          <input type="text" placeholder="Busca" id="campoDeBusca" />
+        </header>
+
+    )
+}
+export default Header
