+++ src/components/InfosUsuario.js
@@ -1,0 +1,12 @@
+function InfosUsuario(){
+    const titulo = "Título do vídeo";
+    const autor = "Joy"
+    return(
+        <h4>
+            {titulo}
+            <br/>
+            {autor}
+        </h4>
+    )
+}
+export default InfosUsuario
