+++ src/components/MenuVerycal.js
@@ -1,0 +1,15 @@
+function MenuVertical (){
+    return(
+       
+            <ul>
+              <li className="botoes-meunu-vertical">Início</li>
+              <li className="botoes-meunu-vertical">Em alta</li>
+              <li className="botoes-meunu-vertical">Inscrições</li>
+              <hr />
+              <li className="botoes-meunu-vertical">Originais</li>
+              <li className="botoes-meunu-vertical">Histórico</li>
+            </ul>
+    )
+
+}
+export default MenuVertical
